make offload-build have pretty colors in makechrootpkg
ssh by default breaks color because it doesn't allocate a pty. We cannot
use ssh -t, because stdin is a pipeline accepting a tarball. Instead,
wrap the build command so it thinks it has a tty even though it doesn't.
All output is going to the other end of an interactive ssh command
anyway.