parse-submodules: update the git submodule command
As of git 2.38.1, the handling of submodules in PKGBUILDS is broken due to CVE-2022-39253.
This will result in "fatal: transport 'file' not allowed" error since
the value of `protocol.file.allow` is changed to be "user" by default.

This commit updates the git submodule command to explicitly use
the file protocol thus fixes the mentioned issue.

Signed-off-by: Orhun Parmaksız <orhunparmaksiz@gmail.com>